pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kittencast"
version = "0.1.0"
description = "A lightweight local EPUB to Audiobook compiler utilizing KittenTTS."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "EbookLib>=0.18",
    "beautifulsoup4>=4.12.0",
    "nltk>=3.8.1",
    "numpy>=1.24.0",
    "soundfile>=0.12.1",
    "tqdm>=4.65.0",
    "pypdf>=4.1.0",
    "python-docx>=1.1.0",
    "kittentts @ https://github.com/KittenML/KittenTTS/releases/download/0.8.1/kittentts-0.8.1-py3-none-any.whl"
]

[project.scripts]
kittencast = "kittencast.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]